+++ data/src/rsbudget/data/api/bo/Plan.java
@@ -3,6 +3,7 @@
  */
 package rsbudget.data.api.bo;
 
+import java.math.BigDecimal;
 import java.util.Set;
 
 import rs.baselib.util.RsMonth;
@@ -39,25 +40,25 @@ public interface Plan extends RsBudgetBO<Long> {
 	 * Returns the balanceStart.
 	 * @return the balanceStart
 	 */
-	public Float getBalanceStart();
+	public BigDecimal getBalanceStart();
 
 	/**
 	 * Sets the balanceStart.
 	 * @param balanceStart the balanceStart to set
 	 */
-	public void setBalanceStart(Float balanceStart);
+	public void setBalanceStart(BigDecimal balanceStart);
 
 	/**
 	 * Returns the balanceEnd.
 	 * @return the balanceEnd
 	 */
-	public Float getBalanceEnd();
+	public BigDecimal getBalanceEnd();
 
 	/**
 	 * Sets the balanceEnd.
 	 * @param balanceEnd the balanceEnd to set
 	 */
-	public void setBalanceEnd(Float balanceEnd);
+	public void setBalanceEnd(BigDecimal balanceEnd);
 
 	/**
 	 * Returns all associated budgets.
